refactor: refactor product-related components and state management
- Update imports for `Product` in `category-preview.component.tsx`, `product-card.component.tsx`, `cart.context.tsx`, `category.component.tsx`
- Remove `ProductsProvider` from `index.tsx`
- Replace `useContext` with `useSelector` in `categories-preview.component.tsx` and `category.component.tsx`
- Add `setCategories` action and `getCategoriesAndDocuments` function in `shop.component.tsx`
