+++ src/routes/categories-preview/categories-preview.component.tsx
@@ -1,9 +1,10 @@
-import { useContext } from 'react';
-import { ProductsContext } from '../../contexts/products.context';
+import { useSelector } from 'react-redux';
 import CategoryPreview from '../../components/category-preview/category-preview.component';
+import { selectCategories } from '../../store/categories/category.selector';
 
 const CategoriesPreview = () => {
-  const { collections } = useContext(ProductsContext);
+  const collections = useSelector(selectCategories);
+
   return (
     <div className='category-preview-container'>
       {collections.map((collection) => {
