+++ src/routes/shop/shop.component.tsx
@@ -1,10 +1,25 @@
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
 import { Routes, Route } from 'react-router-dom';
 import CategoriesPreview from '../categories-preview/categories-preview.component';
 import Category from '../category/category.component';
+import { setCategories } from '../../store/categories/category.action';
 
 import './shop.styles.scss';
+import { getCategoriesAndDocuments } from '../../utils/firebase/firebase.utils';
 
 const Shop = () => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const fetchData = async () => {
+      const categories = await getCategoriesAndDocuments();
+      dispatch(setCategories(categories));
+    };
+
+    fetchData();
+  }, []);
+
   return (
     <Routes>
       <Route index element={<CategoriesPreview />} />
